chore: move cached iterators & wrappers
