+++ pkg/storage/storagewrappers/cached_iterators_test.go
@@ -1,4 +1,4 @@
-package graph
+package storagewrappers
 
 import (
 	"context"
