+++ pkg/storage/storagewrappers/check_caching_test.go
@@ -1,4 +1,4 @@
-package graph
+package storagewrappers
 
 import (
 	"context"
