+++ pkg/storage/storagewrappers/request.go
@@ -3,7 +3,6 @@ package storagewrappers
 import (
 	openfgav1 "github.com/openfga/api/proto/openfga/v1"
 
-	"github.com/openfga/openfga/internal/graph"
 	"github.com/openfga/openfga/internal/server/config"
 	"github.com/openfga/openfga/internal/shared"
 	"github.com/openfga/openfga/pkg/logger"
@@ -26,7 +25,7 @@ func NewRequestStorageWrapperForCheckAPI(ds storage.RelationshipTupleReader, req
 	a = NewBoundedConcurrencyTupleReader(ds, maxConcurrentReads) // to rate-limit reads
 	if cacheSettings.ShouldCacheIterators() {
 		// TODO(miparnisari): pass cacheSettings and resources directly, i can't now because i would create a package import cycle
-		a = graph.NewCachedDatastore(resources.ServerCtx, a, resources.CheckCache, int(cacheSettings.CheckIteratorCacheMaxResults), cacheSettings.CheckIteratorCacheTTL, resources.SingleflightGroup, resources.WaitGroup, []graph.CachedDatastoreOpt{graph.WithCachedDatastoreLogger(logger)}...) // to read tuples from cache
+		a = NewCachedDatastore(resources.ServerCtx, a, resources.CheckCache, int(cacheSettings.CheckIteratorCacheMaxResults), cacheSettings.CheckIteratorCacheTTL, resources.SingleflightGroup, resources.WaitGroup, []CachedDatastoreOpt{WithCachedDatastoreLogger(logger)}...) // to read tuples from cache
 	}
 	b := NewInstrumentedOpenFGAStorage(a)                   // to capture metrics
 	c := NewCombinedTupleReader(b, requestContextualTuples) // to read the contextual tuples
