+++ pkg/storage/storagewrappers/request_test.go
@@ -8,7 +8,6 @@ import (
 
 	openfgav1 "github.com/openfga/api/proto/openfga/v1"
 
-	"github.com/openfga/openfga/internal/graph"
 	"github.com/openfga/openfga/internal/mocks"
 	"github.com/openfga/openfga/internal/server/config"
 	"github.com/openfga/openfga/internal/shared"
@@ -47,7 +46,7 @@ func TestRequestStorageWrapper(t *testing.T) {
 		b, ok := a.RelationshipTupleReader.(*InstrumentedOpenFGAStorage)
 		require.True(t, ok)
 
-		c, ok := b.RelationshipTupleReader.(*graph.CachedDatastore)
+		c, ok := b.RelationshipTupleReader.(*CachedDatastore)
 		// require.Equal(t, mockCache, c.cache)
 		// require.Equal(t, sf, c.sf)
 		// require.Equal(t, 1000, c.maxResultSize)
